requirements.txt
sounddevice
numpy
scipy
openai-whisper